Skip unnecessary pre-commit tasks
Given that A) the pre-commit tasks for black, mypy, and pylint pass the
filenames to the programs and B) not all files are passed to them,
there may be commits where no staged files are touched by them. In
these cases, the pre-commit check can fail because pylint always
expects a file or module. This change ensures that these tasks will
only be run if there is actually anything to check, thereby A) skipping
unnecessary tasks and B) preventing falsely failed pre-commit checks.